Add image

+++ app/controllers/ProductController.php
@@ -1,12 +1,15 @@
 <?php
 
-class ProductController extends \BaseController {
+class ProductController extends BaseController {
 
 	/**
 	 * Display a listing of the resource.
 	 *
 	 * @return Response
 	 */
+	public function __construct() {
+		$this->beforeFilter('csrf', array('on'=>'post'));
+	}
 	public function index()
 	{
 		$products = Prod::All();	
@@ -53,6 +56,10 @@ class ProductController extends \BaseController {
 			$product->price         = Input::get('price');
 			$product->brand_id      = Input::get('brand');
 			$product->category_id   = Input::get('category');
+			$image = Input::file('product_pic');
+			$filename = date('Y-m-d-H:i:s')."-".$image->getClientOriginalName();
+			Image::make($image->getRealPath())->resize(468,249)->save(public_path().'/img/products/'.$filename);
+			$product->product_pic = 'img/products/'.$filename;
 			$product->save();
 
 			// redirect
@@ -120,6 +127,10 @@ class ProductController extends \BaseController {
 			$product->price         = Input::get('price');
 			$product->brand_id      = Input::get('brand');
 			$product->category_id   = Input::get('category');
+			$image = Input::file('product_pic');
+			$filename = date('Y-m-d-H:i:s')."-".$image->getClientOriginalName();
+			Image::make($image->getRealPath())->resize(468,249)->save(public_path().'/img/products/'.$filename);
+			$product->product_pic = 'img/products/'.$filename;
 			$product->save();
 
 			// redirect
@@ -137,10 +148,16 @@ class ProductController extends \BaseController {
 	public function destroy($id)
 	{
 		$product = Prod::find($id);
-		$product->delete();
+		if ($product) {
+			File::delete('public/'.$product->image);
+			$product->delete();
+			Session::flash('message', 'Successfully deleted the product!');
+		return Redirect::to('product');
+		}
+		
 
 		// redirect
-		Session::flash('message', 'Successfully deleted the product!');
+		Session::flash('message', 'Something went wrong, please try again');
 		return Redirect::to('product');
 	}
